Move Select to common
